Removed creation of XML
The XML creation does not really fit into a tool written for
visualisation (except maybe if you want to visualise directly from
*.gene files which does not happen here). If it is required it is better
to put it into an own tool.

As an addition: The tool is no longer linked to libqnn. This should
reduce the footprint. It also makes it possible to run the tool on
systems without libqnn.